pyproject.toml
[tool.poetry]
name = "pfncps"
version = "0.1.0"
description = "Particle Filter Neural Circuit Policies"
authors = ["ten2mushi"]
packages = [
    { include = "pfncps", from = ".." },
]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0"
numpy = "^1.24"
ncps = "^0.0.7"
tqdm = "^4.67.1"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
black = "^23.0"
ruff = "^0.1"
mypy = "^1.0"
hypothesis = {version = ">=6.0.0", python = ">=3.10"}
matplotlib = ">=3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
